backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-magic==0.4.27
alembic==1.13.3
sqladmin==0.20.1
passlib[bcrypt]==1.7.4
pytest==7.4.2
httpx==0.24.1
factory-boy==3.3.0
email-validator==2.0.0
argon2-cffi==21.3.0
python-ulid==3.1.0
Jinja2==3.1.2
PyJWT==2.10.1
cryptography==44.0.0
Pillow==11.1.0
resend==2.7.0
